+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.3.0] - 2024-11-26
+
+- Added `no-std` support
+- Added `hashbrown` as a dependency for `no-std` compatibility
+- Added `no-std` Crate category
+- Fixed tests for well-known inputs
+
 ## [v0.2.2] - 2024-11-25
 
 - Added `BuildHasher` implementation, supporting `HashMap` and `HashSet`
@@ -38,7 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added benchmarks
 - Released a crate
 
-[Unreleased]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.2.2...v0.3.0
 [v0.2.2]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/thevilledev/ChibiHash-rs/compare/v0.1.2...v0.2.0